+++ CHANGELOG.md
@@ -7,6 +7,15 @@ All notable changes to this project will be documented in this file. It uses the
   [Semantic Versioning]: https://semver.org/spec/v2.0.0.html
     "Semantic Versioning 2.0.0"
 
+## [v0.1.4] — Unreleased
+
+### ⚡ Improvements
+
+*   Added support for the PostgreSQL `md5()` function, mapped to
+    `lower(hex(MD5()))` in ClickHouse.
+
+  [v0.1.4]: https://github.com/ClickHouse/pg_clickhouse/compare/v0.1.3...v0.1.4
+
 ## [v0.1.3] — 2026-01-23
 
 This release makes binary-only changes. Once installed, any existing use of